Move dashboard count formatters to a tested lib; delete dead OverviewStat0
formatRawCount/formatCompactCount move to src/lib/format-count.ts with unit
tests locking the compact-rounding edge cases. OverviewStat0 was exported
but imported nowhere.
