+++ src/components/screens/DashboardScreen/OverviewStat.tsx
@@ -5,58 +5,7 @@ import {
   HoverCardContent,
   HoverCardTrigger,
 } from "@/components/ui/hover-card";
-
-/** Full count with grouping separators, e.g. 7434 → "7,434". */
-const formatRawCount = (n: number) => n.toLocaleString("en-US");
-
-/**
- * Abbreviated counts for tight layouts:
- * 7434 → "~7.4k", 10100 → "~10.1k", 101234 → "~101k".
- * Values under 1000 stay as the raw locale string.
- */
-const formatCompactCount = (n: number) => {
-  if (n < 1000) return formatRawCount(n);
-
-  const thousands = n / 1000;
-  if (n >= 100_000) {
-    return `~${Math.round(thousands)}k`;
-  }
-
-  const rounded = Math.round(thousands * 10) / 10;
-  const body = Number.isInteger(rounded) ? String(rounded) : rounded.toFixed(1);
-  return `~${body}k`;
-};
-
-export const OverviewStat0 = ({
-  value,
-  label,
-  Icon,
-  className,
-}: {
-  value: number | string;
-  label: string;
-  Icon: LucideIcon;
-  className?: string;
-}) => (
-  <div
-    className={cn(
-      "flex items-center gap-3 rounded-2xl border-2 border-b-4 border-foreground/15 bg-muted/30 px-3 py-3 text-left",
-      className
-    )}
-  >
-    <div className="flex items-center justify-center border-2 size-9 shrink-0 rounded-xl border-foreground/10 bg-background text-foreground/70">
-      <Icon className="size-4" aria-hidden />
-    </div>
-    <div className="min-w-0">
-      <div className="text-base font-extrabold leading-none tabular-nums">
-        {value}
-      </div>
-      <div className="mt-1 truncate text-[11px] font-bold uppercase tracking-wide text-muted-foreground">
-        {label}
-      </div>
-    </div>
-  </div>
-);
+import { formatCompactCount, formatRawCount } from "@/lib/format-count";
 
 const StatValue = ({
   display,
